Move remaining flake8-pytest-style violations to rule modules
