+++ src/registry.rs
@@ -394,26 +394,26 @@ ruff_macros::define_rule_mapping!(
     PT003 => rules::flake8_pytest_style::rules::ExtraneousScopeFunction,
     PT004 => rules::flake8_pytest_style::rules::MissingFixtureNameUnderscore,
     PT005 => rules::flake8_pytest_style::rules::IncorrectFixtureNameUnderscore,
-    PT006 => violations::ParametrizeNamesWrongType,
-    PT007 => violations::ParametrizeValuesWrongType,
-    PT008 => violations::PatchWithLambda,
-    PT009 => violations::UnittestAssertion,
-    PT010 => violations::RaisesWithoutException,
-    PT011 => violations::RaisesTooBroad,
-    PT012 => violations::RaisesWithMultipleStatements,
-    PT013 => violations::IncorrectPytestImport,
-    PT015 => violations::AssertAlwaysFalse,
-    PT016 => violations::FailWithoutMessage,
-    PT017 => violations::AssertInExcept,
-    PT018 => violations::CompositeAssertion,
+    PT006 => rules::flake8_pytest_style::rules::ParametrizeNamesWrongType,
+    PT007 => rules::flake8_pytest_style::rules::ParametrizeValuesWrongType,
+    PT008 => rules::flake8_pytest_style::rules::PatchWithLambda,
+    PT009 => rules::flake8_pytest_style::rules::UnittestAssertion,
+    PT010 => rules::flake8_pytest_style::rules::RaisesWithoutException,
+    PT011 => rules::flake8_pytest_style::rules::RaisesTooBroad,
+    PT012 => rules::flake8_pytest_style::rules::RaisesWithMultipleStatements,
+    PT013 => rules::flake8_pytest_style::rules::IncorrectPytestImport,
+    PT015 => rules::flake8_pytest_style::rules::AssertAlwaysFalse,
+    PT016 => rules::flake8_pytest_style::rules::FailWithoutMessage,
+    PT017 => rules::flake8_pytest_style::rules::AssertInExcept,
+    PT018 => rules::flake8_pytest_style::rules::CompositeAssertion,
     PT019 => rules::flake8_pytest_style::rules::FixtureParamWithoutValue,
     PT020 => rules::flake8_pytest_style::rules::DeprecatedYieldFixture,
     PT021 => rules::flake8_pytest_style::rules::FixtureFinalizerCallback,
     PT022 => rules::flake8_pytest_style::rules::UselessYieldFixture,
-    PT023 => violations::IncorrectMarkParenthesesStyle,
+    PT023 => rules::flake8_pytest_style::rules::IncorrectMarkParenthesesStyle,
     PT024 => rules::flake8_pytest_style::rules::UnnecessaryAsyncioMarkOnFixture,
     PT025 => rules::flake8_pytest_style::rules::ErroneousUseFixturesOnFixture,
-    PT026 => violations::UseFixturesWithoutParameters,
+    PT026 => rules::flake8_pytest_style::rules::UseFixturesWithoutParameters,
     // flake8-pie
     PIE790 => rules::flake8_pie::rules::NoUnnecessaryPass,
     PIE794 => rules::flake8_pie::rules::DupeClassFieldDefinitions,
