+++ src/rules/flake8_pytest_style/rules/fail.rs
@@ -4,8 +4,20 @@ use super::helpers::{is_empty_or_null_string, is_pytest_fail};
 use crate::ast::helpers::SimpleCallArgs;
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
+use crate::define_violation;
 use crate::registry::Diagnostic;
-use crate::violations;
+use crate::violation::Violation;
+use ruff_macros::derive_message_formats;
+
+define_violation!(
+    pub struct FailWithoutMessage;
+);
+impl Violation for FailWithoutMessage {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("No message passed to `pytest.fail()`")
+    }
+}
 
 pub fn fail_call(checker: &mut Checker, call: &Expr, args: &[Expr], keywords: &[Keyword]) {
     if is_pytest_fail(call, checker) {
@@ -15,13 +27,13 @@ pub fn fail_call(checker: &mut Checker, call: &Expr, args: &[Expr], keywords: &[
         if let Some(msg) = msg {
             if is_empty_or_null_string(msg) {
                 checker.diagnostics.push(Diagnostic::new(
-                    violations::FailWithoutMessage,
+                    FailWithoutMessage,
                     Range::from_located(call),
                 ));
             }
         } else {
             checker.diagnostics.push(Diagnostic::new(
-                violations::FailWithoutMessage,
+                FailWithoutMessage,
                 Range::from_located(call),
             ));
         }
