+++ src/rules/flake8_pytest_style/rules/mod.rs
@@ -1,18 +1,22 @@
 pub use assertion::{
     assert_falsy, assert_in_exception_handler, composite_condition, unittest_assertion,
+    AssertAlwaysFalse, AssertInExcept, CompositeAssertion, UnittestAssertion,
 };
-pub use fail::fail_call;
+pub use fail::{fail_call, FailWithoutMessage};
 pub use fixture::{
     fixture, DeprecatedYieldFixture, ErroneousUseFixturesOnFixture, ExtraneousScopeFunction,
     FixtureFinalizerCallback, FixtureParamWithoutValue, FixturePositionalArgs,
     IncorrectFixtureNameUnderscore, IncorrectFixtureParenthesesStyle, MissingFixtureNameUnderscore,
     UnnecessaryAsyncioMarkOnFixture, UselessYieldFixture,
 };
-pub use imports::{import, import_from};
-pub use marks::marks;
-pub use parametrize::parametrize;
-pub use patch::patch_with_lambda;
-pub use raises::{complex_raises, raises_call};
+pub use imports::{import, import_from, IncorrectPytestImport};
+pub use marks::{marks, IncorrectMarkParenthesesStyle, UseFixturesWithoutParameters};
+pub use parametrize::{parametrize, ParametrizeNamesWrongType, ParametrizeValuesWrongType};
+pub use patch::{patch_with_lambda, PatchWithLambda};
+pub use raises::{
+    complex_raises, raises_call, RaisesTooBroad, RaisesWithMultipleStatements,
+    RaisesWithoutException,
+};
 
 mod assertion;
 mod fail;
