+++ src/violations.rs
@@ -8,9 +8,6 @@ use serde::{Deserialize, Serialize};
 
 use crate::define_violation;
 use crate::rules::flake8_debugger::types::DebuggerUsingType;
-use crate::rules::flake8_pytest_style::types::{
-    ParametrizeNameType, ParametrizeValuesRowType, ParametrizeValuesType,
-};
 use crate::rules::pyupgrade::types::Primitive;
 use crate::violation::{AlwaysAutofixableViolation, AutofixKind, Availability, Violation};
 
@@ -3758,199 +3755,6 @@ impl Violation for DotFormatInException {
     #[derive_message_formats]
     fn message(&self) -> String {
         format!("Exception must not use a `.format()` string directly, assign to variable first")
-    }
-}
-
-// flake8-pytest-style
-
-define_violation!(
-    pub struct ParametrizeNamesWrongType {
-        pub expected: ParametrizeNameType,
-    }
-);
-impl AlwaysAutofixableViolation for ParametrizeNamesWrongType {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let ParametrizeNamesWrongType { expected } = self;
-        format!("Wrong name(s) type in `@pytest.mark.parametrize`, expected `{expected}`")
-    }
-
-    fn autofix_title(&self) -> String {
-        let ParametrizeNamesWrongType { expected } = self;
-        format!("Use a `{expected}` for parameter names")
-    }
-}
-
-define_violation!(
-    pub struct ParametrizeValuesWrongType {
-        pub values: ParametrizeValuesType,
-        pub row: ParametrizeValuesRowType,
-    }
-);
-impl Violation for ParametrizeValuesWrongType {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let ParametrizeValuesWrongType { values, row } = self;
-        format!("Wrong values type in `@pytest.mark.parametrize` expected `{values}` of `{row}`")
-    }
-}
-
-define_violation!(
-    pub struct PatchWithLambda;
-);
-impl Violation for PatchWithLambda {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Use `return_value=` instead of patching with `lambda`")
-    }
-}
-
-define_violation!(
-    pub struct UnittestAssertion {
-        pub assertion: String,
-    }
-);
-impl AlwaysAutofixableViolation for UnittestAssertion {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let UnittestAssertion { assertion } = self;
-        format!("Use a regular `assert` instead of unittest-style `{assertion}`")
-    }
-
-    fn autofix_title(&self) -> String {
-        let UnittestAssertion { assertion } = self;
-        format!("Replace `{assertion}(...)` with `assert ...`")
-    }
-}
-
-define_violation!(
-    pub struct RaisesWithoutException;
-);
-impl Violation for RaisesWithoutException {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("set the expected exception in `pytest.raises()`")
-    }
-}
-
-define_violation!(
-    pub struct RaisesTooBroad {
-        pub exception: String,
-    }
-);
-impl Violation for RaisesTooBroad {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let RaisesTooBroad { exception } = self;
-        format!(
-            "`pytest.raises({exception})` is too broad, set the `match` parameter or use a more \
-             specific exception"
-        )
-    }
-}
-
-define_violation!(
-    pub struct RaisesWithMultipleStatements;
-);
-impl Violation for RaisesWithMultipleStatements {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("`pytest.raises()` block should contain a single simple statement")
-    }
-}
-
-define_violation!(
-    pub struct IncorrectPytestImport;
-);
-impl Violation for IncorrectPytestImport {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Found incorrect import of pytest, use simple `import pytest` instead")
-    }
-}
-
-define_violation!(
-    pub struct AssertAlwaysFalse;
-);
-impl Violation for AssertAlwaysFalse {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Assertion always fails, replace with `pytest.fail()`")
-    }
-}
-
-define_violation!(
-    pub struct FailWithoutMessage;
-);
-impl Violation for FailWithoutMessage {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("No message passed to `pytest.fail()`")
-    }
-}
-
-define_violation!(
-    pub struct AssertInExcept {
-        pub name: String,
-    }
-);
-impl Violation for AssertInExcept {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let AssertInExcept { name } = self;
-        format!(
-            "Found assertion on exception `{name}` in except block, use `pytest.raises()` instead"
-        )
-    }
-}
-
-define_violation!(
-    pub struct CompositeAssertion;
-);
-impl Violation for CompositeAssertion {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Assertion should be broken down into multiple parts")
-    }
-}
-
-define_violation!(
-    pub struct IncorrectMarkParenthesesStyle {
-        pub mark_name: String,
-        pub expected_parens: String,
-        pub actual_parens: String,
-    }
-);
-impl AlwaysAutofixableViolation for IncorrectMarkParenthesesStyle {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let IncorrectMarkParenthesesStyle {
-            mark_name,
-            expected_parens,
-            actual_parens,
-        } = self;
-        format!(
-            "Use `@pytest.mark.{mark_name}{expected_parens}` over \
-             `@pytest.mark.{mark_name}{actual_parens}`"
-        )
-    }
-
-    fn autofix_title(&self) -> String {
-        "Add/remove parentheses".to_string()
-    }
-}
-
-define_violation!(
-    pub struct UseFixturesWithoutParameters;
-);
-impl AlwaysAutofixableViolation for UseFixturesWithoutParameters {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Useless `pytest.mark.usefixtures` without parameters")
-    }
-
-    fn autofix_title(&self) -> String {
-        "Remove `usefixtures` decorator or pass parameters".to_string()
     }
 }
